common-requirements.txt
-e git+https://github.com/django/django.git#egg=Django-dev
psycopg2
south
django_selector
